server/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
psycopg2-binary
uvicorn